Add message parsing
There are a few important bits in each message, but it's a bit hard to
reason about it from the full message dumps themselves.

This commit adds msg_parser.py which:
- detects each message type (M0-M4, BP)
- extracts interesting chunks from the message

This lets me compare changing bits across messages of different
transactions (exchanges) with more scripting and less manual copy-paste.