Fix after-timestamp checksum explanation
There are two bytes after the 6 byte timestamp (read in M1, written in
M3). The second byte is a regular sum-based checksum of the preceeding
14 bytes.

The other byte, now re-labeled x, is not strongly connected to the
checksum (it looks like they sum to around 256, but it's inconsistent).
However, the byte is the same in M3 as in M1, so it can just be copied.